src/rapidapi_toolllm/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-dotenv==1.0.0
httpx==0.25.2
openai==1.3.5
python-jose==3.3.0
aiohttp==3.9.1 